Issue #9554 - add javadoc for huffman / n-bit integer classes and remove static decode methods
